rm `start_simulation`
